Create get_circle_info
This program uses python to pass 5 arguments through a function (get_circle_info()) to determine the point P = (x, y), the center  of a circle C = (h,k) , the distance between points P and C, and the radius of the circle, r. The program then evaluates the point P = (x,y) to determine if this coordinate is on, inside, or outside the graph of the circle through evaluating if and elif statements.